# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dochub API
  version: "1.0.0"
  description: |
    Organizational RAG document hub.

    **Nothing reaches the backend while a user is choosing sources.** The browser holds
    repository details, folder ids and local files until Process is pressed.

    **Process contract.** `POST /artifacts/{artifactId}/process` is the only call the
    Upload screen makes. For each source it writes a `source_documents` row with status
    `RequestUpload` and puts the source on the extract queue
    (`dochub-source-upload-requested`). The extractor service claims one message at a
    time, flips the row to `Uploading`, moves the files to Azure Blob storage under
    `teams/{team}/groups/{group}/artifacts/{artifact}/{uploadedAt}`, sets `Uploaded`, and
    hands the documents to the process queue (`dochub-document-process-requested`). For
    every blob it stores the content MD5 Azure returned in a `document_versions` row.
    The RAG platform (`ragplatform/`) consumes the process queue, indexes each document
    and reports back through the **Ingestion** endpoints: once per document, then once
    for the artifact, which closes the source and sets the artifact's status.

    Local files are the one exception that needs a prior call: their bytes go to
    `POST /uploads/staging` immediately before Process, and the returned staging ids are
    passed in the source's `options`.

    **Recurring updates.** A SharePoint or Google Drive source may carry a `schedule`,
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/uploads/staging:
    post:
      tags: [Processing]
      summary: Stream local file bytes to the server, immediately before pressing Process
      description: |
        Only needed for local files, whose bytes the browser cannot hand to the extractor
        directly. Send each file as a form file and its browser-relative path as a sibling
        text field named `path:{fieldName}`, so folder structure survives into the blob
        prefix. Nothing is recorded against an artifact until Process is called.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file0: { type: string, format: binary }
                "path:file0": { type: string }
      responses:
        "200":
          description: Staged
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StagedFile" } }
        "413": { $ref: "#/components/responses/Error" }
  /api/artifacts/{artifactId}/process:
    parameters:
      - { name: artifactId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Processing]
      summary: Submit one or more sources for processing
      description: |
        The only call the Upload screen makes. Every source is validated before any of
        them is written, so a bad entry in a batch cannot leave half the submission queued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sources]
              properties:
                sources:
                  type: array
                  minItems: 1
                  items: { $ref: "#/components/schemas/SourceSubmission" }
      responses:
        "202":
          description: Queued — poll each source's status
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources: { type: array, items: { $ref: "#/components/schemas/SourceDocument" } }
                  schedules: { type: array, items: { $ref: "#/components/schemas/SyncSchedule" } }
                  message: { type: string, example: "SRC-1042 queued · 4 documents from Claims IT Technical. We'll notify you as it progresses." }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/source-documents:
    get:
      tags: [Processing]
      summary: The processing list for the active organization
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { name: artifactId, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SourceDocument" } }
  /api/source-documents/{sourceDocumentId}:
    parameters:
      - { name: sourceDocumentId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Processing]
      summary: One submitted source with its per-document detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  source: { $ref: "#/components/schemas/SourceDocument" }
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
        "404": { $ref: "#/components/responses/Error" }
  /api/source-documents/{sourceDocumentId}/retry:
    parameters:
      - { name: sourceDocumentId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Processing]
      summary: Put a failed source back on the extract queue
      responses:
        "202": { description: Re-queued }
        "409":
          description: Only a failed source can be retried
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `POST /api/auth/sso`. Carries the active `org_id`, so every request is tenant-scoped by the token.
    serviceKey:
      type: apiKey
      in: header
      name: X-Dochub-Service-Key
      description: Equal to `Ingestion:ServiceKey` (≥32 chars). Used only by the RAG platform.
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: owner_only }
        message: { type: string }
        details: { nullable: true }
    SourceType:
      type: string
      enum: [Local, GitHub, SharePoint, GoogleDrive, AzureDevOps, Confluence, Jira]
    SourceDocumentStatus:
      type: string
      description: |
        The whole journey of one submitted source.
        RequestUpload → Uploading → Uploaded → Processing → Processed
      enum: [RequestUpload, Uploading, Uploaded, Processing, Processed, PartiallyFailed, Failed, Cancelled]
    DocumentStatus:
      type: string
      description: State of a single file inside a source.
      enum: [Pending, Uploading, Uploaded, Processing, Indexed, Failed, Skipped]
    SourceSubmission:
      type: object
      required: [sourceType, sourceReference]
      properties:
        sourceType: { $ref: "#/components/schemas/SourceType" }
        sourceReference:
          type: string
          description: Human-readable pointer shown in the UI.
          example: acme-ins/claims-core@main:/docs
        sourceConnectionId:
          type: string
          format: uuid
          nullable: true
          description: Required for SharePoint and Google Drive.
        options:
          type: object
          description: |
            Connector payload. GitHub: `repository`, `branch`, `path`, `fileTypes`.
            SharePoint: `driveId`, `itemId`. Google Drive: `folderId`.
            Azure DevOps: `organization`, `project`, `wiki`. Local: `stagingIds`.
          additionalProperties: true
          example: { repository: "acme-ins/claims-core", branch: "main", path: "docs", fileTypes: "md, yaml, adoc" }
        documents:
          type: array
          nullable: true
          description: Pre-declared files, so rows exist before extraction begins. Local sources only.
          items:
            type: object
            required: [name, relativePath]
            properties:
              name: { type: string }
              relativePath: { type: string }
              sizeBytes: { type: integer, format: int64, nullable: true }
              contentType: { type: string, nullable: true }
        schedule:
          nullable: true
          description: Omit for a one-time import; supply it to also keep the location in sync.
          allOf: [{ $ref: "#/components/schemas/SyncScheduleRequest" }]
    SourceDocument:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference: { type: string, example: SRC-1042 }
        artifactId: { type: string, format: uuid }
        artifactName: { type: string }
        path: { type: string, example: "Claims › Claims IT" }
        sourceType: { $ref: "#/components/schemas/SourceType" }
        sourceReference: { type: string }
        status: { $ref: "#/components/schemas/SourceDocumentStatus" }
        statusLabel: { type: string, example: Request Upload }
        totalDocuments: { type: integer }
        uploadedDocuments: { type: integer }
        processedDocuments: { type: integer }
        failedDocuments: { type: integer }
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
          description: Spans the whole journey — uploading fills the first half, processing the second.
        blobContainer: { type: string, nullable: true }
        blobPrefix:
          type: string
          nullable: true
          example: teams/claims/groups/claims-it/artifacts/claims-it-technical/20260927T164512123Z
        requestedBy: { type: string }
        isScheduled: { type: boolean, description: True when a recurring schedule submitted this. }
        requestedAt: { type: string, format: date-time }
        uploadStartedAt: { type: string, format: date-time, nullable: true }
        uploadedAt: { type: string, format: date-time, nullable: true }
        processingStartedAt: { type: string, format: date-time, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
    StagedFile:
      type: object
      properties:
        stagingId: { type: string }
        name: { type: string }
        relativePath: { type: string, nullable: true }
        sizeBytes: { type: integer, format: int64 }
        contentType: { type: string, nullable: true }
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        sourceDocumentId: { type: string, format: uuid }
        name: { type: string }
        relativePath: { type: string }
        sourceLocation: { type: string }
        sourceType: { $ref: "#/components/schemas/SourceType" }
        status: { $ref: "#/components/schemas/DocumentStatus" }
        sizeBytes: { type: integer, format: int64 }
        contentType: { type: string, nullable: true }
        blobPath:
          type: string
          nullable: true
          example: teams/claims/groups/claims-it/artifacts/claims-it-technical/20260925T164512123Z/Data Dictionary v4.docx
        blobUrl: { type: string, nullable: true }
        blobUploadedAt: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
        revision:
          type: integer
          description: Incremented each time a recurring sync replaces this document's content.
        lastSyncedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        contentMd5:
          type: string
          nullable: true
          description: Base64 MD5 that Azure Blob returned for the current version.
          example: Cc1DVWbsE3AuUE4jDeOrGQ==
        chunkCount: { type: integer, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
...
    SyncScheduleRequest:
      type: object
      required: [frequency, timeOfDay]
      properties:
        frequency: { type: string, enum: [Daily, Weekly, Monthly] }
        timeOfDay:
          type: string
          description: Wall-clock time in `timeZoneId`, 24-hour.
          example: "09:00"
        timeZoneId:
          type: string
          description: IANA id. Defaults to UTC, so a cadence survives DST changes intact.
          example: Europe/London
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
          nullable: true
          description: 0 = Sunday. Weekly only.
        dayOfMonth:
          type: integer
          minimum: 1
          maximum: 28
          nullable: true
          description: Capped at 28 so the run happens in every month. Monthly only.
    SyncSchedule:
      type: object
      properties:
        id: { type: string, format: uuid }
        artifactId: { type: string, format: uuid }
        artifactName: { type: string }
        path: { type: string, example: "Policy › Policy Admin" }
        sourceType: { type: string, enum: [SharePoint, GoogleDrive] }
        sourceReference: { type: string }
        frequency: { type: string, enum: [Daily, Weekly, Monthly] }
        timeOfDay: { type: string, example: "09:00" }
        timeZoneId: { type: string, example: Europe/London }
        dayOfWeek: { type: integer, nullable: true }
        dayOfMonth: { type: integer, nullable: true }
        status: { type: string, enum: [Active, Paused, Failed] }
        cadence: { type: string, example: Weekly on Monday at 09:00 Europe/London }
        nextRunAt: { type: string, format: date-time }
        lastRunAt: { type: string, format: date-time, nullable: true }
        documentsAddedLastRun: { type: integer }
        documentsUpdatedLastRun: { type: integer }
        documentsUnchangedLastRun: { type: integer }
        consecutiveFailures: { type: integer, description: After 5 the schedule pauses itself and notifies. }
        lastError: { type: string, nullable: true }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
...